ALX SE

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -13,7 +13,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 	new = malloc(sizeof(hash_table_t));
 	if (new == NULL)
 		return (NULL);
-	if (size < 0)
+	if ((int)size < 0)
 		return (NULL);
 	new->array = malloc(sizeof(hash_node_t *) * size);
 	if (new->array == NULL)
